style(tests,tools): apply the mechanical clang-tidy readability fixes
Behaviour-preserving fixes surfaced by widening the lint scope (#296):
uppercase integer-literal suffixes, explicit parentheses around mixed
* / + arithmetic, one declaration per statement, redundant parentheses and
duplicate prototypes removed, concise #ifdef/#elifdef, widening casts moved
onto the operands of a multiplication, and the s_ prefix on file- and
function-scope statics.

Three renames are hand-made because the auto-fix would have picked the
wrong answer: the jpeg segment fixtures are file-scope statics and take
s_, not the g_ that GlobalConstantPrefix would apply; t_png_cursor becomes
png_cursor_t; and the streamed-JPEG local renamed to s_src would have
collided with the file-scope s_src byte buffer, so it is s_pull_src.

Names fixed by an external ABI keep their spelling under a per-site NOLINT
that states why: _sbrk (newlib), _GNU_SOURCE (glibc feature test), and the
six mock-interposition seams whose macro name must equal the production
symbol it replaces -- renaming one silently un-hooks the mock and the test
would pass against the real driver.

Two real defects fell out: board_sim called cs_disasm with sizeof(code)
on an array parameter, handing capstone the pointer width instead of the
four valid instruction bytes; and tools/vela/vela_gen.py emitted an
unprefixed file-scope static, so the generator and its committed output
both now emit s_.

Refs #296

+++ tests/support/fs_fat_exfat_write_test_util.h
@@ -179,7 +179,9 @@ static inline ra8_err_t wc_read(void* ctx, uint32_t lba, uint32_t count, uint8_t
   if (lba + count > d->block_count) {
     return k_ra8_err_out_of_range;
   }
-  memcpy(buf, &d->bytes[lba * (uint32_t)k_wc_block_size], (size_t)count * (size_t)k_wc_block_size);
+  memcpy(buf,
+         &d->bytes[(size_t)lba * (uint32_t)k_wc_block_size],
+         (size_t)count * (size_t)k_wc_block_size);
   return k_ra8_ok;
 }
 
@@ -212,7 +214,9 @@ static inline ra8_err_t wc_write(void* ctx, uint32_t lba, uint32_t count, const 
   if (lba + count > d->block_count) {
     return k_ra8_err_out_of_range;
   }
-  memcpy(&d->bytes[lba * (uint32_t)k_wc_block_size], buf, (size_t)count * (size_t)k_wc_block_size);
+  memcpy(&d->bytes[(size_t)lba * (uint32_t)k_wc_block_size],
+         buf,
+         (size_t)count * (size_t)k_wc_block_size);
   return k_ra8_ok;
 }
 
@@ -314,8 +318,8 @@ static inline void build_exfat_volume(void)
  */
 static inline uint32_t root_entry_off(const ra8_fs_mount_t* h, uint32_t idx)
 {
-  const uint32_t root_lba = h->first_data_lba + (h->root_cluster - 2U) * h->sectors_per_cluster;
-  return root_lba * (uint32_t)k_wc_block_size + idx * (uint32_t)k_wc_entry_bytes;
+  const uint32_t root_lba = h->first_data_lba + ((h->root_cluster - 2U) * h->sectors_per_cluster);
+  return (root_lba * (uint32_t)k_wc_block_size) + (idx * (uint32_t)k_wc_entry_bytes);
 }
 
 /**
@@ -333,7 +337,7 @@ static inline uint32_t root_entry_off(const ra8_fs_mount_t* h, uint32_t idx)
  */
 static inline uint32_t fat_entry_off(const ra8_fs_mount_t* h, uint32_t clus)
 {
-  return h->first_fat_lba * (uint32_t)k_wc_block_size + clus * 4U;
+  return (h->first_fat_lba * (uint32_t)k_wc_block_size) + (clus * 4U);
 }
 
 /**
